+++ test/search-spec.js
@@ -3,7 +3,7 @@ var request = require('supertest')(config.baseUrl);
 var cheerio = require('cheerio');
 
 describe('Search Page', function () {
-    it('Default Page', function (done) {
+    it('default Page', function (done) {
         request.get('/search')
             .expect(200)
             .expect(function (res) {
@@ -14,7 +14,7 @@ describe('Search Page', function () {
             .end(done);
     });
 
-    it('Search article', function (done) {
+    it('search article', function (done) {
         request.post('/search')
             .send({
                 'searchContent': '1',
@@ -28,5 +28,21 @@ describe('Search Page', function () {
                 }
             })
             .end(done);
-    })
+    });
+
+    it('search user', function (done) {
+        request.post('/search')
+            .send({
+                'searchContent': '小',
+                'searchType': 'author'
+            })
+            .expect(200)
+            .expect(function (res) {
+                var $ = cheerio.load(res.text);
+                if ($("#user tbody tr td").eq(0).text().length < 1) {
+                    throw new Error("author name must exist.");
+                }
+            })
+            .end(done);
+    });
 });